fastapi/requirements.txt
fastapi
uvicorn
ffmpeg-python
python-multipart
openai-whisper